python_backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
numpy==2.1.3
pillow==11.3.0
scikit-learn==1.7.0
tensorflow==2.19.1
gunicorn==21.2.0
tensorflow==2.19.1